requirements.txt
fastapi>=0.115.0,<1.0
uvicorn[standard]>=0.32.0
jinja2>=3.1.0
python-multipart>=0.0.12
python-dotenv>=1.0.0
httpx>=0.27.0
supabase>=2.0.0
gspread>=6.0.0
google-auth>=2.25.0
google-api-python-client>=2.100.0
google-generativeai>=0.8.0
openai>=1.0.0
yt-dlp>=2024.0.0
sse-starlette>=1.8,<2.0
